server: split webhook handlers by forge provider
